# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "127.0.0.1:8000/"
        },
        {
            "url": "127.0.0.1:8000/"
        },
        {
            "url": "127.0.0.1:8000/"
        }
    ],
    "paths": {
...
        "/api/ajouter-role": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Ajouter un role",
                "description": "Ajoute un nouveau role avec les détails fournis",
                "operationId": "ajouterRole",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nomRole": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ajout role  avec succès"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/ajoutercandidat": {
            "post": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Inscrire un candidat ",
                "description": "Inscription d'un candidat  avec les détails fournis",
                "operationId": "inscrirCandidat",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "prenom": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "adresse": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Candidat inscrit avec succès"
                    }
                }
            }
        },
        "/api/ajouter-utilisateur-admin": {
            "post": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Inscrire un admin",
                "description": "Inscription d'un admin  avec les détails fournis",
                "operationId": "ajouterUtilisateurAdmin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "prenom": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Admin inscrit avec succès"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Connecter un utilisateur",
                "description": "Connexion d'un utilisteurs  avec les détails fournis",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur connecté avec succès"
                    },
                    "401": {
                        "description": "Informations invalid"
                    }
                }
            }
        },
        "/api/deconnecter": {
            "post": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Deconnecter un utilisateur",
                "description": "Deconnexion d'un utilisteurs  et invalidation de son token jwt",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Utilisateur deconnecté  avec succès"
                    },
                    "401": {
                        "description": "Aucun utilisateur connecté"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }
    },
...
}